secured post deleting embed post form

+++ app/controllers/posts_controller.rb
@@ -2,6 +2,7 @@ class PostsController < ApplicationController
 
   def index
   	@posts = Post.all
+    @posts = @posts.reverse
     @user = current_user
   end
 
@@ -18,20 +19,29 @@ class PostsController < ApplicationController
   end
 
   def create
+    session[:return_to] ||= request.referer
     @post = Post.new(happ_params)
     @post.user_id = current_user.id;
    
     @post.save
-    redirect_to @post
+    redirect_to session.delete(:return_to)
   end
 
   def delete
+    if current_user.nil?
+      redirect_to root_path
+    else
       session[:return_to] ||= request.referer
       @posts = Post.all
       @post = Post.find(params[:id])
-      @post.destroy
-      #redirect_to root_path
-      redirect_to session.delete(:return_to)
+      if current_user.id != @post.user_id
+        redirect_to root_path
+      else
+        @post.destroy
+        #redirect_to root_path
+        redirect_to session.delete(:return_to)
+      end
+    end
   end
 
   private
